protect default roles from modification through api calls
