+++ src/main/java/org/folio/roles/controller/RoleCapabilitySetController.java
@@ -12,6 +12,7 @@ import org.folio.roles.rest.resource.RoleCapabilitySetApi;
 import org.folio.roles.service.capability.CapabilitySetService;
 import org.folio.roles.service.capability.RoleCapabilitySetService;
 import org.folio.roles.service.role.RoleService;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -21,6 +22,7 @@ public class RoleCapabilitySetController implements RoleCapabilitySetApi {
 
   private final RoleService roleService;
   private final CapabilitySetService capabilitySetService;
+  @Qualifier("apiRoleCapabilitySetService")
   private final RoleCapabilitySetService roleCapabilitySetService;
 
   @Override
