+++ src/main/java/org/folio/roles/service/capability/RoleCapabilitySetService.java
@@ -1,71 +1,20 @@
 package org.folio.roles.service.capability;
 
-import static java.lang.Integer.MAX_VALUE;
-import static java.util.Collections.emptyList;
-import static org.apache.commons.collections4.CollectionUtils.isEmpty;
-import static org.apache.commons.collections4.CollectionUtils.isNotEmpty;
-import static org.folio.common.utils.CollectionUtils.mapItems;
-import static org.folio.roles.domain.entity.RoleCapabilitySetEntity.DEFAULT_ROLE_CAPABILITY_SET_SORT;
-import static org.folio.roles.utils.CapabilityUtils.getCapabilityEndpoints;
-
-import jakarta.persistence.EntityExistsException;
-import jakarta.persistence.EntityNotFoundException;
-import java.util.Collection;
 import java.util.List;
 import java.util.UUID;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.log4j.Log4j2;
-import org.folio.roles.domain.dto.Endpoint;
 import org.folio.roles.domain.dto.RoleCapabilitySet;
-import org.folio.roles.domain.entity.RoleCapabilitySetEntity;
-import org.folio.roles.domain.entity.key.RoleCapabilitySetKey;
 import org.folio.roles.domain.model.PageResult;
-import org.folio.roles.mapper.entity.RoleCapabilitySetEntityMapper;
-import org.folio.roles.repository.RoleCapabilitySetRepository;
-import org.folio.roles.service.permission.RolePermissionService;
-import org.folio.roles.service.role.RoleService;
-import org.folio.roles.utils.UpdateOperationHelper;
-import org.folio.spring.data.OffsetRequest;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
-@Log4j2
-@Service
-@Transactional
-@RequiredArgsConstructor
-public class RoleCapabilitySetService {
-
-  private final RoleService roleService;
-  private final CapabilityService capabilityService;
-  private final CapabilitySetService capabilitySetService;
-  private final RolePermissionService rolePermissionService;
-  private final CapabilityEndpointService capabilityEndpointService;
-  private final RoleCapabilitySetRepository roleCapabilitySetRepository;
-  private final RoleCapabilitySetEntityMapper roleCapabilitySetEntityMapper;
+public interface RoleCapabilitySetService {
 
   /**
    * Creates a record(s) associating one or more capabilitySets with a role.
    *
    * @param roleId - role identifier as {@link UUID} object
    * @param capabilitySetIds - capabilitySet identifiers as {@link List} of {@link UUID} objects
-   * @return {@link PageResult} with created {@link RoleCapabilitySet} relations
+   * @return {@link PageResult} with created {@link org.folio.roles.domain.dto.RoleCapabilitySet} relations
    */
-  @Transactional
-  public PageResult<RoleCapabilitySet> create(UUID roleId, List<UUID> capabilitySetIds) {
-    if (isEmpty(capabilitySetIds)) {
-      throw new IllegalArgumentException("List with capability set identifiers is empty");
-    }
-
-    roleService.getById(roleId);
-    var existingEntities = roleCapabilitySetRepository.findRoleCapabilitySets(roleId, capabilitySetIds);
-    var existingCapabilitySetIds = getSetIds(existingEntities);
-    if (isNotEmpty(existingCapabilitySetIds)) {
-      throw new EntityExistsException(String.format(
-        "Relation already exists for role='%s' and capabilitySets=%s", roleId, existingCapabilitySetIds));
-    }
-
-    return assignCapabilities(roleId, capabilitySetIds, emptyList());
-  }
+  PageResult<RoleCapabilitySet> create(UUID roleId, List<UUID> capabilitySetIds);
 
   /**
    * Retrieves role-capabilitySets items by CQL query.
@@ -73,15 +22,9 @@ public class RoleCapabilitySetService {
    * @param query - CQL query as {@link String} object
    * @param limit - a number of results in response
    * @param offset - offset in pagination from first record.
-   * @return {@link PageResult} with found {@link RoleCapabilitySet} relations
+   * @return {@link PageResult} with found {@link org.folio.roles.domain.dto.RoleCapabilitySet} relations
    */
-  @Transactional(readOnly = true)
-  public PageResult<RoleCapabilitySet> find(String query, Integer limit, Integer offset) {
-    var offsetRequest = OffsetRequest.of(offset, limit, DEFAULT_ROLE_CAPABILITY_SET_SORT);
-    var entities = roleCapabilitySetRepository.findByQuery(query, offsetRequest);
-    var roleCapabilities = entities.map(roleCapabilitySetEntityMapper::convert);
-    return PageResult.of(roleCapabilities.getTotalElements(), roleCapabilities.getContent());
-  }
+  PageResult<RoleCapabilitySet> find(String query, Integer limit, Integer offset);
 
   /**
    * Updates a list of assigned to a role capabilitySets.
@@ -89,16 +32,7 @@ public class RoleCapabilitySetService {
    * @param roleId - role identifier
    * @param capabilityIds - list with new capabilitySets, that should be assigned to a role
    */
-  @Transactional
-  public void update(UUID roleId, List<UUID> capabilityIds) {
-    roleService.getById(roleId);
-    var assignedRoleCapabilityEntities = roleCapabilitySetRepository.findAllByRoleId(roleId);
-
-    var assignedCapabilityIds = getSetIds(assignedRoleCapabilityEntities);
-    UpdateOperationHelper.create(assignedCapabilityIds, capabilityIds, "role-capability set")
-      .consumeAndCacheNewEntities(newIds -> getSetIds(assignCapabilities(roleId, newIds, assignedCapabilityIds)))
-      .consumeDeprecatedEntities((deprecatedIds, createdIds) -> removeCapabilities(roleId, deprecatedIds, createdIds));
-  }
+  void update(UUID roleId, List<UUID> capabilityIds);
 
   /**
    * Removes role assigned capabilities using role identifier.
@@ -106,71 +40,14 @@ public class RoleCapabilitySetService {
    * @param roleId - role identifier as {@link UUID}
    * @throws jakarta.persistence.EntityNotFoundException if role is not found by id or there is no assigned values
    */
-  @Transactional
-  public void deleteAll(UUID roleId) {
-    roleService.getById(roleId);
-    var roleCapabilityEntities = roleCapabilitySetRepository.findAllByRoleId(roleId);
-    if (isEmpty(roleCapabilityEntities)) {
-      throw new EntityNotFoundException("Relations between role and capability sets are not found for role: " + roleId);
-    }
-
-    var capabilitySetIds = getSetIds(roleCapabilityEntities);
-    removeCapabilities(roleId, capabilitySetIds, emptyList());
-  }
+  void deleteAll(UUID roleId);
 
   /**
    * Removes role assigned capability set using role identifier and capability set id.
    *
-   * @param roleId  - role identifier as {@link UUID}
-   * @param capabilitySetId  - capability set identifier as {@link UUID}
+   * @param roleId - role identifier as {@link UUID}
+   * @param capabilitySetId - capability set identifier as {@link UUID}
    * @throws jakarta.persistence.EntityNotFoundException if assignment is not found by role id and capability set id
    */
-  @Transactional
-  public void delete(UUID roleId, UUID capabilitySetId) {
-    var entity = roleCapabilitySetRepository.findById(RoleCapabilitySetKey.of(roleId, capabilitySetId));
-    if (entity.isEmpty()) {
-      throw new EntityNotFoundException("Relations between role and capability set is not found: roleId = " + roleId
-        + ", capabilitySetId = " + capabilitySetId);
-    }
-
-    removeCapabilities(roleId, List.of(capabilitySetId), emptyList());
-  }
-
-  private PageResult<RoleCapabilitySet> assignCapabilities(
-    UUID roleId, List<UUID> newSetIds, Collection<UUID> assignedSetIds) {
-    log.debug("Assigning capabilities to role: roleId = {}, ids = {}", roleId, newSetIds);
-    capabilitySetService.checkIds(newSetIds);
-
-    var entities = mapItems(newSetIds, capabilitySetId -> new RoleCapabilitySetEntity(roleId, capabilitySetId));
-    var changedEndpoints = getChangedEndpoints(roleId, newSetIds, assignedSetIds);
-    rolePermissionService.createPermissions(roleId, changedEndpoints);
-
-    var resultEntities = roleCapabilitySetRepository.saveAll(entities);
-    var createdRoleCapabilities = mapItems(resultEntities, roleCapabilitySetEntityMapper::convert);
-    log.info("Capabilities assigned to role: roleId = {}, ids = {}", roleId, newSetIds);
-
-    return PageResult.of(createdRoleCapabilities.size(), createdRoleCapabilities);
-  }
-
-  private void removeCapabilities(UUID roleId, List<UUID> deprecatedSetIds, Collection<UUID> assignedSetIds) {
-    log.debug("Revoking capabilities from role: roleId = {}, ids = {}", roleId, deprecatedSetIds);
-    var changedEndpoints = getChangedEndpoints(roleId, deprecatedSetIds, assignedSetIds);
-    rolePermissionService.deletePermissions(roleId, changedEndpoints);
-    roleCapabilitySetRepository.deleteRoleCapabilitySets(roleId, deprecatedSetIds);
-    log.info("Capability sets are revoked to role: roleId = {}, ids = {}", roleId, deprecatedSetIds);
-  }
-
-  private List<Endpoint> getChangedEndpoints(UUID roleId, List<UUID> deprecatedIds, Collection<UUID> assignedIds) {
-    var directlyAssignedCapabilities = capabilityService.findByRoleId(roleId, false, MAX_VALUE, 0);
-    var excludedEndpoints = getCapabilityEndpoints(directlyAssignedCapabilities.getRecords());
-    return capabilityEndpointService.getByCapabilitySetIds(deprecatedIds, assignedIds, excludedEndpoints);
-  }
-
-  private static List<UUID> getSetIds(PageResult<RoleCapabilitySet> roleCapabilitySets) {
-    return mapItems(roleCapabilitySets.getRecords(), RoleCapabilitySet::getCapabilitySetId);
-  }
-
-  private static List<UUID> getSetIds(List<RoleCapabilitySetEntity> existingEntities) {
-    return mapItems(existingEntities, RoleCapabilitySetEntity::getCapabilitySetId);
-  }
+  void delete(UUID roleId, UUID capabilitySetId);
 }
